Add streaming Stack ingestion pipeline
